refactor: split large ml_analysis files into subdirectory modules
Split 3 large files into modular subdirectories:

1. convergence.rs (1166 lines) → convergence/ (8 files, max 240 lines)
   - mod.rs (113) - Main API
   - epoch.rs (33) - Epoch progression
   - loss.rs (169) - Loss convergence
   - patterns.rs (91) - Advanced patterns
   - stability.rs (220) - Training stability
   - learning_curves.rs (240) - Learning curve analysis
   - plateau.rs (142) - Plateau detection
   - optimization.rs (231) - Optimization trajectory

2. quantization.rs (871 lines) → quantization/ (5 files, max 325 lines)
   - mod.rs (53) - Main API
   - types.rs (58) - Shared structures
   - precision.rs (325) - Precision analysis
   - methods.rs (185) - Methods analysis
   - impact.rs (269) - Impact analysis

3. gradient.rs (695 lines) → gradient/ (6 files, max 299 lines)
   - mod.rs (54) - Main API
   - types.rs (25) - Shared structures
   - statistics.rs (299) - Statistics extraction
   - magnitudes.rs (65) - Magnitude analysis
   - distributions.rs (59) - Distribution analysis
   - flow.rs (264) - Flow analysis

Result: All source files now under 600 lines (largest: diff.rs 553 lines)
Build verified successful with cargo build --release
